refactor: PascalCase enum members across the engine
Renames screaming-case / abbreviated enum members to idiomatic C#
PascalCase and updates every call site, test, example, and benchmark:
  FileReadCode  OK->Ok, FAIL->Fail, NO_FILE->NoFile
  CullMode      NONE->None, CW->ClockWise, CCW->CounterClockWise
  KeyPress      Ctrl->Control (and related)
Also fixes a stale <see cref> in PlatformOpenTK that still referenced
the old KeyPress.Ctrl name (CS1574 under TreatWarningsAsErrors).

+++ PixelSharper.Core/Extensions/RayCastWorld.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 using PixelSharper.Core.Components;
 using PixelSharper.Core.Types;
 
@@ -322,7 +320,7 @@ public abstract class Engine : PGEX
 
         for (var x = 0; x < _screenSize.X; x++)
         {
-            var rayAngle = _cameraHeading - _fieldOfView / 2f + (float)x / _floatScreenSize.X * _fieldOfView;
+            var rayAngle = _cameraHeading - _fieldOfView / 2f + x / _floatScreenSize.X * _fieldOfView;
             var rayDirX = MathF.Cos(rayAngle);
             var rayDirY = MathF.Sin(rayAngle);
